requirements.txt
# BTC期货自动交易系统最小依赖

# 核心交易库
ccxt>=4.5.1

# Redis客户端
redis>=5.0.0

# YAML配置文件解析
pyyaml>=6.0.0